Changed from gradle to maven multi module project.

+++ library-service/src/main/resources/db/migration/V1__create_book_table.sql
@@ -1,4 +1,4 @@
 CREATE TABLE book (
-    id integer primary key not null,
+    id bigserial primary key not null,
     name varchar(255) not null
 );
